MagenticUI/prompt-http-server/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
httpx==0.28.1
python-dotenv==1.2.1
